+++ layers/Schema/packages/customposts-wp/src/TypeAPIs/CustomPostTypeAPI.php
@@ -172,6 +172,11 @@ class CustomPostTypeAPI extends AbstractCustomPostTypeAPI
             // This param can either be a string or an array. Eg:
             // $query['orderby'] => array('date' => 'DESC', 'title' => 'ASC');
         }
+        // Post slug
+        if (isset($query['slug'])) {
+            $query['name'] = $query['slug'];
+            unset($query['slug']);
+        }
         if (isset($query['post-not-in'])) {
             $query['post__not_in'] = $query['post-not-in'];
             unset($query['post-not-in']);