+++ layers/Schema/packages/post-categories/src/FieldResolvers/RootPostCategoryFieldResolver.php
@@ -22,21 +22,11 @@ class RootPostCategoryFieldResolver extends AbstractQueryableFieldResolver
         return array(RootTypeResolver::class);
     }
 
-    public function getSchemaFieldDescription(TypeResolverInterface $typeResolver, string $fieldName): ?string
-    {
-        $descriptions = [
-            'postCategory' => $this->translationAPI->__('Post category with a specific ID', 'post-categories'),
-            'postCategories' => $this->translationAPI->__('Post categories', 'post-categories'),
-            'postCategoryCount' => $this->translationAPI->__('Number of post categories', 'post-categories'),
-            'postCategoryNames' => $this->translationAPI->__('Names of the post categories', 'post-categories'),
-        ];
-        return $descriptions[$fieldName] ?? parent::getSchemaFieldDescription($typeResolver, $fieldName);
-    }
-
     public function getFieldNamesToResolve(): array
     {
         return [
             'postCategory',
+            'postCategoryBySlug',
             'postCategories',
             'postCategoryCount',
             'postCategoryNames',
@@ -47,6 +37,7 @@ class RootPostCategoryFieldResolver extends AbstractQueryableFieldResolver
     {
         $types = [
             'postCategory' => SchemaDefinition::TYPE_ID,
+            'postCategoryBySlug' => SchemaDefinition::TYPE_ID,
             'postCategories' => SchemaDefinition::TYPE_ID,
             'postCategoryCount' => SchemaDefinition::TYPE_INT,
             'postCategoryNames' => SchemaDefinition::TYPE_STRING,
@@ -67,6 +58,18 @@ class RootPostCategoryFieldResolver extends AbstractQueryableFieldResolver
         };
     }
 
+    public function getSchemaFieldDescription(TypeResolverInterface $typeResolver, string $fieldName): ?string
+    {
+        $descriptions = [
+            'postCategory' => $this->translationAPI->__('Post category with a specific ID', 'post-categories'),
+            'postCategoryBySlug' => $this->translationAPI->__('Post category with a specific slug', 'post-categories'),
+            'postCategories' => $this->translationAPI->__('Post categories', 'post-categories'),
+            'postCategoryCount' => $this->translationAPI->__('Number of post categories', 'post-categories'),
+            'postCategoryNames' => $this->translationAPI->__('Names of the post categories', 'post-categories'),
+        ];
+        return $descriptions[$fieldName] ?? parent::getSchemaFieldDescription($typeResolver, $fieldName);
+    }
+
     public function getSchemaFieldArgs(TypeResolverInterface $typeResolver, string $fieldName): array
     {
         $schemaFieldArgs = parent::getSchemaFieldArgs($typeResolver, $fieldName);
@@ -79,6 +82,18 @@ class RootPostCategoryFieldResolver extends AbstractQueryableFieldResolver
                             SchemaDefinition::ARGNAME_NAME => 'id',
                             SchemaDefinition::ARGNAME_TYPE => SchemaDefinition::TYPE_ID,
                             SchemaDefinition::ARGNAME_DESCRIPTION => $this->translationAPI->__('The category ID', 'post-categories'),
+                            SchemaDefinition::ARGNAME_MANDATORY => true,
+                        ],
+                    ]
+                );
+            case 'postCategoryBySlug':
+                return array_merge(
+                    $schemaFieldArgs,
+                    [
+                        [
+                            SchemaDefinition::ARGNAME_NAME => 'slug',
+                            SchemaDefinition::ARGNAME_TYPE => SchemaDefinition::TYPE_STRING,
+                            SchemaDefinition::ARGNAME_DESCRIPTION => $this->translationAPI->__('The category slug', 'post-categories'),
                             SchemaDefinition::ARGNAME_MANDATORY => true,
                         ],
                     ]
@@ -134,9 +149,13 @@ class RootPostCategoryFieldResolver extends AbstractQueryableFieldResolver
         $postCategoryTypeAPI = PostCategoryTypeAPIFacade::getInstance();
         switch ($fieldName) {
             case 'postCategory':
-                $query = [
-                    'include' => [$fieldArgs['id']],
-                ];
+            case 'postCategoryBySlug':
+                $query = [];
+                if ($fieldName == 'postCategory') {
+                    $query['include'] = [$fieldArgs['id']];
+                } elseif ($fieldName == 'postCategoryBySlug') {
+                    $query['slugs'] = [$fieldArgs['slug']];
+                }
                 $options = [
                     'return-type' => ReturnTypes::IDS,
                 ];
@@ -168,6 +187,7 @@ class RootPostCategoryFieldResolver extends AbstractQueryableFieldResolver
     {
         switch ($fieldName) {
             case 'postCategory':
+            case 'postCategoryBySlug':
             case 'postCategories':
                 return PostCategoryTypeResolver::class;
         }